usracct: added "object" field
There are applications where an authentication is performed to access a
specific object, such as a filename. This is added to the usracct schema.

The hierarchy to describe "what is accessed" becomes:

An authentication is made from "device" to access a given
"object" by using a given "service" implemented by the "application".